requirements.txt.txt
streamlit
scikit-learn
scikit-image
joblib
numpy
matplotlib
Pillow